backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.19
anthropic>=0.30.0
python-dotenv>=1.0.0
